Fix type data of id and use mapper

+++ src/main/java/com/myproject/movie/controllers/UserController.java
@@ -41,7 +41,7 @@ public class UserController {
 
     @PreAuthorize("hasAnyRole('ADMIN', 'USER')")
     @GetMapping("/{id}")
-    public ResponseEntity<UserReadResponseDTO> getUserById(@PathVariable Long id) {
+    public ResponseEntity<UserReadResponseDTO> getUserById(@PathVariable Integer id) {
         log.info("Fetching user with ID: {}", id);
 
         return ResponseEntity.ok(userService.getUserById(id));
@@ -49,7 +49,7 @@ public class UserController {
 
     @PreAuthorize("hasAnyRole('USER', 'ADMIN') and @customPermissionChecker.hasAccess(authentication, #id)")
     @PutMapping("/{id}")
-    public ResponseEntity<UserUpdateResponseDTO> updateUser(@PathVariable Long id, @RequestBody UserUpdateResponseDTO userUpdateResponseDTO) {
+    public ResponseEntity<UserUpdateResponseDTO> updateUser(@PathVariable Integer id, @RequestBody UserUpdateResponseDTO userUpdateResponseDTO) {
         log.info("Updating user with ID: {}", id);
         UserUpdateResponseDTO updatedUser = userService.updateUser(id, userUpdateResponseDTO);
         log.info("User with ID {} updated successfully", id);
@@ -58,7 +58,7 @@ public class UserController {
 
     @PreAuthorize("hasAnyRole('USER', 'ADMIN') and @customPermissionChecker.hasAccess(authentication, #id)")
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteUser(@PathVariable Integer id) {
         log.info("Deleting user with ID: {}", id);
         userService.deleteUser(id);
         log.info("User with ID {} deleted successfully", id);
